+++ scrapy/signals.py
@@ -17,4 +17,8 @@ response_downloaded = object()
 item_scraped = object()
 item_dropped = object()
 
+stats_spider_opened = spider_opened
+stats_spider_closing = spider_closed
+stats_spider_closed = spider_closed
+
 item_passed = item_scraped # for backwards compatibility